+++ R/tidyverse.R
@@ -27,14 +27,10 @@
 #'   using SUPPORT.md.
 #'
 #' * `use_tidy_coc()`: creates code of conduct same as `use_code_of_conduct()`
-#'   but for the fact that it's placed in a .github subdirectory.
+#'   but for the fact that it's placed in a `.github/` subdirectory.
 #'
-#' * `use_tidy_community()`: wrapper function around contributing, issues,
+#' * `use_tidy_github()`: grouping function around contributing, issues,
 #'   support, and code of conduct functions.
-#'
-#' * `use_tidy_github()`: creates standard tidyverse issue template,
-#'   support resources document, contributing guidelines, and code of conduct in
-#'   a `.github/` subdirectory.
 #'
 #' @name tidyverse
 NULL
@@ -201,50 +197,9 @@ use_tidy_coc <- function() {
 
 #' @export
 #' @rdname tidyverse
-use_tidy_community <- function() {
+use_tidy_github <- function() {
   use_tidy_contributing()
   use_tidy_issue_template()
   use_tidy_support()
   use_tidy_coc()
-}
-
-
-#' @export
-#' @rdname tidyverse
-use_tidy_github <- function() {
-  check_uses_github()
-
-  gh <- gh::gh_tree_remote(proj_get())
-  travis_url <- file.path("https://travis-ci.org", gh$username, gh$repo)
-  github_url <- file.path("https://github.com", gh$username, gh$repo)
-
-  use_directory(".github", ignore = TRUE)
-  use_template(
-    "tidy_contributing.md",
-    ".github/CONTRIBUTING.md",
-    data = list(
-      package = project_name(),
-      github_url = github_url,
-      travis_url = travis_url
-    )
-  )
-  use_template(
-    "ISSUE_TEMPLATE.md",
-    ".github/ISSUE_TEMPLATE.md"
-  )
-  use_template(
-    "SUPPORT.md",
-    ".github/SUPPORT.md",
-    data = list(package = project_name())
-  )
-  use_template(
-    "CODE_OF_CONDUCT.md",
-    ".github/CODE_OF_CONDUCT.md"
-  )
-
-  todo("Don't forget to describe the code of conduct in your README.md:")
-  code_block(
-    "Please note that this project is released with a [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md).",
-    "By participating in this project you agree to abide by its terms."
-  )
 }